requirements.txt
google-cloud-firestore==2.11.1
streamlit-extras==0.2.7
Pillow==9.3.0
pandas==1.5.1